fix admin

+++ application/controllers/controller_admin.php
@@ -22,21 +22,36 @@ class Controller_Admin extends Controller
 
    function action_delete()
    {
-      if(!empty($_POST['id'])){
-         include ("database/db.php");
-         mysqli_query($db, "DELETE FROM `users` WHERE id={$_POST['id']}");
-         Header("Location: /admin");
+      if ($_SESSION['admin']==1) {
+         if(!empty($_POST['id'])){
+            include ("database/db.php");
+            mysqli_query($db, "DELETE FROM `users` WHERE id={$_POST['id']}");
+            Header("Location: /admin");
+         }
+      }
+      else {
+         exit("Не все так просто");
       }
    }
 
    function action_editor() 
    {
+      if ($_SESSION['admin']==1) {
       $data = $this->model->get_user();		
 		$this->view->generate('editor_view.php', 'template_view.php', $data);
+      }
+      else {
+         exit("Не все так просто");
+      }
    }
    
    function action_save_user() 
    {
+      if ($_SESSION['admin']==1) {
       $this->model->save_user();
+      }
+      else {
+         exit("Не все так просто");
+      }
    }
 }
